* src/plugin-common.c (plugin_print_stats): remove         \t that was putting garbage in syslog logfile.
        * src/common.c (prelude_log_set_prefix):
        (standard_log):
        (syslog_log):
        new function that allow to set a global prefix to all
        prelude_log() call. Passing NULL to prelude_log_set_prefix()
        remove the prefix.

        * src/daemonize.c (prelude_daemonize): renamed to
        prelude_daemonize(). Comment the function.

        Also, if requested, create an exclusive lockfile held
        until program exit.

        * src/prelude-message.c (read_message_content):
        comment why we return prelude_msg_error if we get an
        EOF in the middle of a message.
        (prelude_msg_read): free message on error / eof.

        * src/prelude-client.c (handle_plaintext_connection):
        (handle_ssl_connection):
        (start_inet_connection):
        (start_unix_connection):
        fit API change.

        * src/prelude-client-mgr.c (setup_backup_fd):

        * src/prelude-io.c (prelude_io_set_file_io): now
        correspond to libc buffered I/O.
        (prelude_io_set_sys_io): renamed from set_network_io().


git-svn-id: file:///home/yoann/SVN/trunk/libprelude@440 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/include/plugin-common-prv.h
@@ -25,6 +25,7 @@
 #define _PLUGIN_COMMON_PRV_H
 
 #include "list.h"
+#include <inttypes.h>
 
 typedef struct 
 {
@@ -51,7 +52,7 @@ typedef struct
          * Plugin running time and count.
          */
         double p_time;
-        unsigned long p_count;
+        uint32_t p_count;
 
         /*
          * Pointer on the parent of this container (on the entry in the lib).
